Replace morgan with winston
- Remove the morgan logger
- Add winston and express-winston

While morgan allows to easily log the routes being hit, it's quite
useful to have debug log messages to help understand how the application
is being used.

Winston gives a logger instance that we can use across the entire
application, while express-winston allows to replace morgan entirely
with the addition of not allowing our application to hard-fail (if we so
desire).

Winston also allows us to change the transport (save logs to file or
stdout) and format messages according to the environment (either easy to
read or JSON)